+++ src/azure-cli/HISTORY.rst
@@ -9,6 +9,7 @@ Release History
 **Appservice**
 
 * functionapp: `az functionapp create` enables application insights by default
+* BREAKING CHANGE: (functionapp) removes deprecated `az functionapp devops-build` command. Please use the new command `az functionapp devops-pipeline` instead.
 
 **DLS**